fixed #922

+++ core/sharedarea.c
@@ -292,7 +292,7 @@ struct uwsgi_sharedarea *uwsgi_sharedarea_init_fd(int fd, uint64_t len, off_t of
 
 struct uwsgi_sharedarea *uwsgi_sharedarea_init(int pages) {
 	int id = uwsgi_sharedarea_new_id();
-	uwsgi.sharedareas[id] = uwsgi_calloc_shared(uwsgi.page_size * (pages + 1));
+	uwsgi.sharedareas[id] = uwsgi_calloc_shared(uwsgi.page_size * (size_t)(pages + 1));
 	uwsgi.sharedareas[id]->area = ((char *) uwsgi.sharedareas[id]) + uwsgi.page_size;
 	uwsgi.sharedareas[id]->id = id;
 	uwsgi.sharedareas[id]->fd = -1;
